refactor : 불필요한 라이브러리 제거
