+++ front/src/pages/competitiondetail/competitiondetailnotice/noticewindow/NoticeWindow.tsx
@@ -1,7 +1,6 @@
-import React, { useState } from 'react'
-import { useQuery, useMutation } from 'react-query'
+import { useState } from 'react'
+import { useMutation } from 'react-query'
 import { useParams, Link } from 'react-router-dom'
-import { getTeamsinLeader } from '@/services/member'
 import ConfirmModal from '@/components/ConfirmModal/ConfirmModal'
 import Button from '@/components/Button/Button'
 import { registerNotice } from '@/services/notice'
