--- front/src/pages/home/HomeOri.module.scss
@@ -1,103 +1,0 @@
-@import '@/styles/variables.scss';
-
-.home {
-  .main {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    font-family: 'Pretendard-Regular';
-    background-image: url('../../../public/images/home/home.webp');
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-
-    .nav {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      width: 100%;
-      margin-top: 2rem;
-      padding: 0 2rem; // Adjust padding as necessary
-    }
-
-    .title {
-      font-size: 3rem;
-      color: white;
-      text-align: center;
-      margin-top: 6rem;
-      margin-bottom: 3rem;
-    }
-
-    .mainDescription {
-      font-size: 20px;
-      color: black;
-      text-align: center;
-      margin-bottom: 8rem;
-      line-height: 1.6;
-    }
-  }
-
-  .contents {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin-top: 4rem;
-    margin-bottom: 4rem;
-    margin-right: 12rem;
-    margin-left: 12rem;
-
-    .contentContainer {
-      display: flex;
-      justify-content: start;
-      align-items: stretch;
-      width: 100%;
-      margin-bottom: 4rem;
-
-      .photo {
-        width: 50%;
-        height: 100%;
-      }
-
-      .content {
-        flex-grow: 1;
-        display: flex;
-        padding-top: 8rem;
-        flex-direction: column;
-        justify-content: start;
-        align-items: stretch;
-
-        .contentTitle {
-          font-size: 2rem;
-          font-weight: bold;
-          margin-bottom: 1rem;
-          border-bottom: 2px solid black;
-        }
-
-        .contentDescription {
-          font-size: 1.2rem;
-        }
-      }
-
-      :nth-child(even).content {
-        align-items: flex-start;
-        text-align: left;
-        margin-left: 4rem;
-      }
-
-      :nth-child(odd).content {
-        align-items: flex-end;
-        text-align: right;
-        margin-right: 4rem;
-      }
-    }
-  }
-}
-
-@font-face {
-  font-family: 'Pretendard-Regular';
-  src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff') format('woff');
-  font-weight: 400;
-  font-style: normal;
-}
