+++ front/src/pages/mypage/myinfo/MyInfoPage.tsx
@@ -1,5 +1,5 @@
 import Navbar from '@/components/Navbar/Navbar'
-import MyPageTab from '@/pages/mypage/MyPageTab'
+import MyPageTab from '@/pages/mypage/tab/Tab'
 import MyInfo from '@/pages/mypage/myinfo/MyInfo'
 import Footer from '@/components/Footer/Footer'
 import styles from '@/pages/mypage/MyPage.module.scss'
