+++ front/src/pages/mypage/teamlist/TeamListPage.tsx
@@ -1,5 +1,5 @@
 import Navbar from '@/components/Navbar/Navbar'
-import MyPageTab from '@/pages/mypage/MyPageTab'
+import MyPageTab from '@/pages/mypage/tab/Tab'
 import styles from '@/pages/mypage/MyPage.module.scss'
 import TeamList from '@/pages/mypage/teamlist/TeamList'
 import Footer from '@/components/Footer/Footer'
